requirements.txt
uvicorn
fastapi
openai-whisper
ollama